CMake: Migrate to CMake-target based linking
Linking against libraries specified in CMake variables is not the modern
way of linking in CMake. Rather, linking against CMake targets is
preferred as it is no longer required to set include dirs, compile flags
etc. separately.

This commit removes calls to target_include_directories() with CMake
variables and changes the used CMake variables in
target_link_libraries() calls to their respective CMake targets.

Furthermore, the find modules are updated to create imported targets to
be used as link targets later on.

Signed-off-by: NexAdn <nex@nexadn.de>